// src/types/Car.ts
export type Car = {
  id: number;
  plate: string;
  manufacture: string;
  image: string;
  model: string;
  type: string;
  description: string;
  transmission: string;
  capacity: number;
  rentPerDay: number;
  availableAt: string;
  available: boolean;
  year: number;
  options: string[];
  specs: string[];
};